Bugfix: Mapping in cpwer was in special cases suboptimal
Those hyp/ref, that got no ref/hyp assigned where ignored in the
estimation of the best mapping. Hence, it could happen (typically with
bad transcriptions), that the estimated mapping is not the best mapping.
Now the full cost_matrix is estimated and for over- and underestimation,
the mapping is also correct, when the transcriptions are bad.

Bug found by Samuele Cornell.