Improve rendering of search matches
Change defaultl color to a bright yellow. This is closer to the commonly used color for search results and ensures that the text is still readable.

Only color the outline with the highlight color in the details plot so that the match color is still visible when zoomed in far enough