Fix edge case where MIMO-WER differs from tcMIMO-WER
Edge case: two reference utterances don't overlap temporally, but they both overlap with the same word from the hypothesis. Before this change, tcMIMO-WER did not allow the two utterances to change their order. Now it does.

This introduces potential for a long run-time in extreme cases, but it is unlikely to be a problem in practice.